Spinning out `entity::Enemy` module and moving initialization code there
